requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==3.7.1
attrs==25.4.0
billiard==4.2.2
cachetools==6.2.1
celery==5.3.4
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coloredlogs==15.0.1
distro==1.9.0
feedparser==6.0.12
flatbuffers==25.9.23
google-ai-generativelanguage==0.4.0
google-api-core==2.26.0
google-auth==2.41.1
google-generativeai==0.3.0
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.62.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
idna==3.11
imageio==2.37.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
lazy_loader==0.4
llvmlite==0.45.1
mpmath==1.3.0
networkx==3.5
newsapi-python==0.2.7
numba==0.62.1
numpy==2.2.6
onnxruntime==1.23.1
openai>=1.54.0
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==11.3.0
platformdirs==4.5.0
pooch==1.8.2
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==4.25.8
psycopg2-binary==2.9.11
pyaes==1.6.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.2
pydantic_core==2.41.4
PyMatting==1.1.14
pyTelegramBotAPI==4.14.0
python-dateutil==2.8.2
python-dotenv==1.0.0
redis==5.0.1
referencing==0.37.0
rembg==2.0.67
replicate==0.15.0
requests==2.31.0
rpds-py==0.27.1
rsa==4.9.1
scikit-image==0.25.2
scipy==1.16.2
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
Telethon==1.32.0
tifffile==2025.10.4
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
